+++ ZawodyWin/ZawodyWin/DataModels/Tournament.cs
@@ -1,19 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.Serialization;
 
 namespace ZawodyWin.DataModels
 {
 
     public class Tournament
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public DateTime? Date { get; set; }
-        public int? OrganizerId { get; set; }
+        public long? OrganizerId { get; set; }
         public string? Place { get; set; }
-        public int? LeadingRefereeId { get; set; }
+        public long? LeadingRefereeId { get; set; }
     }
 }
